+++ src-docs/src/theme_k6_dark.scss
@@ -1,2 +1,3 @@
 @import "../../src/theme_k6_dark";
 @import "./components/guide_components";
+@import "./views/header/global_filter_group";
