+++ src-docs/src/theme_light.scss
@@ -1,3 +1,4 @@
 @import "../../src/theme_light";
 @import "./components/guide_components";
+@import "./views/header/global_filter_group";
