+++ src-docs/src/views/header/_global_filter_bar.scss
@@ -1,0 +1,18 @@
+.globalFilterBar {
+  // Match just the regular drop shadow of inputs
+  @include euiFormControlDefaultShadow;
+  display: flex;
+  align-items: center;
+  padding: 4px;
+  min-height: $euiFormControlHeight;
+  margin: 0;
+}
+
+.globalFilterBar__pinned {
+  padding: $euiSizeXS;
+  background-color: saturate(tintOrShade($euiColorSecondary, 40%, 50%), 50%);
+}
+
+.globalFilterBar__pinnedIcon {
+  margin: -2px 0;
+}
