+++ src-docs/src/views/header/_global_filter_item.scss
@@ -1,0 +1,12 @@
+.globalFilterItem {
+
+}
+
+.globalFilterItem-isDisabled {
+  font-weight: $euiFontWeightRegular;
+  font-style: italic;
+}
+
+.globalFilterItem-isExcluded {
+  text-decoration: line-through;
+}
